don't query with unresolved references
